+++ src-tauri/src/config.rs
@@ -21,6 +21,7 @@ pub struct Config {
     pub start_minimized: bool,
     pub enable_multiplayer_hubs: bool,
     pub enable_vatsim_traffic: bool,
+    pub inject_butterlog_traffic: bool,
 }
 
 impl Config {
@@ -59,6 +60,7 @@ impl Config {
             start_minimized: false,
             enable_multiplayer_hubs: false,
             enable_vatsim_traffic: false,
+            inject_butterlog_traffic: false,
         }
     }
 }
@@ -79,6 +81,7 @@ impl Default for Config {
             start_minimized: false,
             enable_multiplayer_hubs: false,
             enable_vatsim_traffic: false,
+            inject_butterlog_traffic: false,
         }
     }
 }